Correcciones ultima entrega
+getValor en Indicador que busca en el repo de IndicadoresConValor

+CargarCuentasEmpresas mejorada para no procesar el mismo archivo dos
veces (lo mueve a otro directorio con la fecha que fue procesado)

+++ src/main/java/application/cargacuentas/CargaCuentasEmpresas.java
@@ -1,29 +1,31 @@
 package application.cargacuentas;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.SocketException;
+import java.time.LocalDateTime;
+import java.util.Arrays;
 import java.util.List;
 
 import org.apache.commons.net.ftp.FTPClient;
 
+import exceptions.ArchivoConErroresException;
 import model.Empresa;
 import model.data.HandlerArchivoJSON;
 import model.repositories.RepoEmpresas;
 
 public class CargaCuentasEmpresas {
 
-	private static FTPClient ftp;
+	private FTPClient ftp;
 
-	private static String rutaFTP = "/CuentasEmpresas.json";
+	private String rutaFTP = "CuentasEmpresas.json";
 
-	private static String ip = "jpbulbulian.brickftp.com";
-	private static String user = "grupo12";
-	private static String pass = "grupo12grupo12";
+	private String ip = "jpbulbulian.brickftp.com";
+	private String user = "grupo12";
+	private String pass = "grupo12grupo12";
 
-	public static void conectar(String ip, String user, String pass) throws SocketException, IOException {
+	public void conectar(String ip, String user, String pass) throws SocketException, IOException {
 		ftp = new FTPClient();
 		ftp.connect(ip);
 
@@ -33,32 +35,66 @@ public class CargaCuentasEmpresas {
 			System.out.println("login Error");
 	}
 
-	public static File descargarArchivoFTP(String hostFile)
-			throws FileNotFoundException, IOException {
-		File tmpFile = File.createTempFile("CuentasEmpresas", ".json");
-		FileOutputStream fos = new FileOutputStream(tmpFile);
-		if (ftp.retrieveFile(hostFile, fos))
-			System.out.println("Descarga correcta");
-		else
-			System.out.println("Error Descarga");
+	public boolean descargar(File dest) throws IOException {
+		FileOutputStream fos = new FileOutputStream(dest);
+		boolean r;
+		if (ftp.retrieveFile(rutaFTP, fos)){
+			System.out.println("descarga OK");
+			r = true;
+		}
+		else{
+			System.out.println("descarga Error");
+			r = false;
+		}
 		fos.close();
-		return tmpFile;
+		return r;
 	}
 
-	private static void desconectar() throws IOException {
+	private void desconectar() throws IOException {
 		ftp.disconnect();
 	}
 
+	private void moveTo(String directory) throws IOException{
+		if(ftp.cwd(directory) == 550)
+			ftp.makeDirectory("/" + directory);
+		else
+			ftp.cwd("../");
+		String[] newName = rutaFTP.split("\\.(?=[^\\.]+$)",2);
+		newName[0] += ("-" + directory.toUpperCase() + "-" + LocalDateTime.now());
+		ftp.rename(rutaFTP, directory + "/" + newName[0] + "." + newName[1]);
+	}
+	
+	private void moveToProcessed() throws IOException{
+		moveTo("Processed");
+	}
+	
+	private void moveToError() throws IOException{
+		moveTo("Error");
+	}
+	
 	public void cargar(){
 		try {
 			conectar(ip, user, pass);
-			File tmpFile = descargarArchivoFTP(rutaFTP);
-			List<Empresa> empresas = new HandlerArchivoJSON(tmpFile.getAbsolutePath()).loadEmpresas();
-			tmpFile.delete();
-			RepoEmpresas.getInstance().insertarVarios(empresas);
-		} catch (IOException e) {
+			if(Arrays.stream(ftp.listNames()).anyMatch(s -> s.equals(rutaFTP))){
+				File tmpFile = File.createTempFile("CuentasEmpresas", ".json");
+				if(descargar(tmpFile)){
+					try{
+						List<Empresa> empresas = new HandlerArchivoJSON(tmpFile.getAbsolutePath()).loadEmpresas();
+						RepoEmpresas.getInstance().insertarVarios(empresas);
+					}
+					catch(ArchivoConErroresException e){
+						moveToError();
+					}
+				}
+				tmpFile.delete();
+				moveToProcessed();
+			}
+//			ftp.rename("Processed/CuentasEmpresas-PROCESSED-2017-11-23T19:11:20.154.json", rutaFTP);
+		}
+		catch (Exception e){
 			e.printStackTrace();
-		} finally{
+		}
+		finally{
 			try {
 				desconectar();
 			} catch (IOException e) {
